+++ src/main/java/com/oracle/dragon/util/exception/OCIDatabaseWaitForShutdownFailedException.java
@@ -1,0 +1,7 @@
+package com.oracle.dragon.util.exception;
+
+public class OCIDatabaseWaitForShutdownFailedException extends DSException {
+    public OCIDatabaseWaitForShutdownFailedException(Exception e) {
+        super(ErrorCode.OCIDatabaseWaitForShutdownFailed,"Unable to wait for complete shutdown of the database!",e);
+    }
+}