+++ src/main/java/com/oracle/dragon/util/exception/OCIDatabaseWaitForStartFailedException.java
@@ -1,0 +1,7 @@
+package com.oracle.dragon.util.exception;
+
+public class OCIDatabaseWaitForStartFailedException extends DSException {
+    public OCIDatabaseWaitForStartFailedException(Exception e) {
+        super(ErrorCode.OCIDatabaseWaitForStartFailed,"Unable to wait for complete startup of the database!",e);
+    }
+}